requirements.txt
asgiref==3.5.2
autopep8==1.7.0
aws-psycopg2==1.3.8
backports.zoneinfo==0.2.1
black==22.10.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
coverage==6.5.0
coveralls==3.3.1
cryptography==38.0.1
defusedxml==0.7.1
Django==4.1.2
django-allauth==0.51.0
docopt==0.6.2
gunicorn==20.1.0
idna==3.4
mypy-extensions==0.4.3
numpy==1.23.4
oauthlib==3.2.2
pathspec==0.10.1
platformdirs==2.5.2
pycodestyle==2.9.1
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.4
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
typing-extensions==4.4.0
urllib3==1.26.12
